requirements.txt
Pygments~=2.10.0
Pillow~=8.3.2